Added way to render to texture - still have not yet tested. Implemented a very light and simple OpenGL bind state mirror layer to (hopefully) prevent redundant state changes

+++ Graphics/Font.cpp
@@ -469,7 +469,7 @@ void setGlyphTexCoords(const Glyph& g, const FontMap& fm, Quad& q)
 {
 	Rectui r = {
 		uint32_t(g.xOffsetCoord),
-		fm.height - g.bitmap.h,
+		uint32_t(fm.height - g.bitmap.h),
 		uint32_t(g.bitmap.w),
 		uint32_t(g.bitmap.h)
 	};
